promote status module to directory
